Benford's law for first and second digits

First and second digit probabilities, as postulated by Benford's law.

theta_benford(d)

Arguments

d
Numeric vector of length 1. Admissible values: 1 or 2.

Value

theta_benford(1) returns a numeric vector of length 9.

theta_benford(2) returns a numeric vector length 10.

Details

Benford’s law is an empirical regularity regarding the occurrence of digits in numbers that was first introduced by Newcomb (1881) and later popularized and supported with empirical evidence by Benford (1938) . The formulas that give Benford's law first and second digit probabilities can be consulted in Hill (1995) .

References

Benford F (1938). “The law of anomalous numbers.” Proceedings of the American Philosophical Society, 78(4).

Hill TP (1995). “A statistical derivation of the significant-digit law.” Statistical science, 10(4), 354--363.

Newcomb S (1881). “Note on the frequency of use of the different digits in natural numbers.” American Journal of mathematics, 4(1), 39--40.

See also

Examples

theta_benford(1)

#> [1] 0.30103000 0.17609126 0.12493874 0.09691001 0.07918125 0.06694679 0.05799195
#> [8] 0.05115252 0.04575749

theta_benford(2)

#>  [1] 0.11967927 0.11389010 0.10882150 0.10432956 0.10030820 0.09667724
#>  [7] 0.09337474 0.09035199 0.08757005 0.08499735